+++ packages/studio/src/hooks/useDomEditSession.ts
@@ -1,53 +1,16 @@
-import { useCallback, useRef, useEffect } from "react";
+import { useEffect } from "react";
 import type { TimelineElement } from "../player";
-import { usePlayerStore } from "../player";
-import { FONT_EXT } from "../utils/mediaTypes";
-import { applyPatchByTarget, type PatchOperation } from "../utils/sourcePatcher";
-import { saveProjectFilesWithHistory } from "../utils/studioFileHistory";
-import {
-  confirmElementDelete,
-  isImageBackgroundValue,
-  isManualGeometryStyleProperty,
-  normalizeDomEditStyleValue,
-  type RightPanelTab,
-} from "../utils/studioHelpers";
-import {
-  primaryFontFamilyValue,
-  injectPreviewGoogleFont,
-  injectPreviewImportedFont,
-  ensureImportedFontFace,
-} from "../utils/studioFontHelpers";
 import { STUDIO_INSPECTOR_PANELS_ENABLED } from "../components/editor/manualEditingAvailability";
-import {
-  buildDomEditStylePatchOperation,
-  buildDomEditTextPatchOperation,
-  findElementForSelection,
-  getDomEditTargetKey,
-  isTextEditableSelection,
-  serializeDomEditTextFields,
-  buildDefaultDomEditTextField,
-  type DomEditTextField,
-  type DomEditSelection,
-} from "../components/editor/domEditing";
-import {
-  removeStudioManualEditsForSelection,
-  type StudioManualEditManifest,
-  upsertStudioBoxSizeEdit,
-  upsertStudioPathOffsetEdit,
-  upsertStudioRotationEdit,
-} from "../components/editor/manualEdits";
-import {
-  removeStudioMotionForSelection,
-  type StudioGsapMotion,
-  type StudioMotionManifest,
-  upsertStudioGsapMotion,
-} from "../components/editor/studioMotion";
-import { fontFamilyFromAssetPath, type ImportedFontAsset } from "../components/editor/fontAssets";
-import type { DomEditGroupPathOffsetCommit } from "../components/editor/DomEditOverlay";
+import { findElementForSelection } from "../components/editor/domEditing";
+import type { StudioManualEditManifest } from "../components/editor/manualEdits";
+import type { StudioMotionManifest } from "../components/editor/studioMotion";
+import type { ImportedFontAsset } from "../components/editor/fontAssets";
 import type { EditHistoryKind } from "../utils/editHistory";
+import type { RightPanelTab } from "../utils/studioHelpers";
 import { useAskAgentModal } from "./useAskAgentModal";
 import { useDomSelection } from "./useDomSelection";
 import { usePreviewInteraction } from "./usePreviewInteraction";
+import { useDomEditCommits } from "./useDomEditCommits";
 
 // ── Types ──
 
@@ -221,495 +184,52 @@ export function useDomEditSession({
     setAgentModalOpen,
   });
 
-  // ── Refs ──
+  // ── Commit handlers (delegated to useDomEditCommits) ──
 
-  const domTextCommitVersionRef = useRef(0);
-
-  // ── Callbacks ──
-
-  const resolveImportedFontAsset = useCallback(
-    (fontFamilyValue: string): ImportedFontAsset | null => {
-      const family = primaryFontFamilyValue(fontFamilyValue);
-      if (!family) return null;
-      const imported = importedFontAssetsRef.current.find(
-        (font) => font.family.toLowerCase() === family.toLowerCase(),
-      );
-      if (imported) return imported;
-      const asset = fileTree.find(
-        (path) =>
-          FONT_EXT.test(path) &&
-          fontFamilyFromAssetPath(path).toLowerCase() === family.toLowerCase(),
-      );
-      if (!asset) return null;
-      return {
-        family: fontFamilyFromAssetPath(asset),
-        path: asset,
-        url: `/api/projects/${projectId}/preview/${asset}`,
-      };
-    },
-    [fileTree, projectId, importedFontAssetsRef],
-  );
-
-  const persistDomEditOperations = useCallback(
-    async (
-      selection: DomEditSelection,
-      operations: Parameters<typeof applyPatchByTarget>[2][],
-      options?: {
-        label?: string;
-        coalesceKey?: string;
-        skipRefresh?: boolean;
-        prepareContent?: (html: string, sourceFile: string) => string;
-        shouldSave?: () => boolean;
-      },
-    ) => {
-      const pid = projectIdRef.current;
-      if (!pid) throw new Error("No active project");
-      if (options?.shouldSave && !options.shouldSave()) return;
-
-      const targetPath = selection.sourceFile || activeCompPath || "index.html";
-      const response = await fetch(`/api/projects/${pid}/files/${encodeURIComponent(targetPath)}`);
-      if (!response.ok) {
-        throw new Error(`Failed to read ${targetPath}`);
-      }
-
-      const data = (await response.json()) as { content?: string };
-      const originalContent = data.content;
-      if (typeof originalContent !== "string") {
-        throw new Error(`Missing file contents for ${targetPath}`);
-      }
-
-      let patchedContent = originalContent;
-      for (const operation of operations) {
-        patchedContent = applyPatchByTarget(patchedContent, selection, operation);
-      }
-      if (options?.prepareContent) {
-        patchedContent = options.prepareContent(patchedContent, targetPath);
-      }
-      if (options?.shouldSave && !options.shouldSave()) return;
-
-      if (patchedContent === originalContent) {
-        throw new Error(`Unable to patch ${selection.selector ?? selection.id ?? "selection"}`);
-      }
-
-      await saveProjectFilesWithHistory({
-        projectId: pid,
-        label: options?.label ?? "Edit layer",
-        kind: "manual",
-        coalesceKey: options?.coalesceKey,
-        files: { [targetPath]: patchedContent },
-        readFile: async () => originalContent,
-        writeFile: writeProjectFile,
-        recordEdit: editHistory.recordEdit,
-      });
-
-      if (options?.skipRefresh) {
-        domEditSaveTimestampRef.current = Date.now();
-      } else {
-        setRefreshKey((k) => k + 1);
-      }
-    },
-    [
-      activeCompPath,
-      editHistory.recordEdit,
-      writeProjectFile,
-      projectIdRef,
-      domEditSaveTimestampRef,
-      setRefreshKey,
-    ],
-  );
-
-  const handleDomPathOffsetCommit = useCallback(
-    (selection: DomEditSelection, next: { x: number; y: number }) => {
-      commitStudioManualEditManifestOptimistically(
-        (manifest) => upsertStudioPathOffsetEdit(manifest, selection, next),
-        {
-          label: "Move layer",
-          coalesceKey: `path-offset:${getDomEditTargetKey(selection)}`,
-        },
-      );
-      refreshDomEditSelectionFromPreview(selection);
-    },
-    [commitStudioManualEditManifestOptimistically, refreshDomEditSelectionFromPreview],
-  );
-
-  const handleDomGroupPathOffsetCommit = useCallback(
-    (updates: DomEditGroupPathOffsetCommit[]) => {
-      if (updates.length === 0) return;
-      const coalesceKey = updates
-        .map((update) => getDomEditTargetKey(update.selection))
-        .sort()
-        .join(":");
-      commitStudioManualEditManifestOptimistically(
-        (manifest) =>
-          updates.reduce(
-            (nextManifest, update) =>
-              upsertStudioPathOffsetEdit(nextManifest, update.selection, update.next),
-            manifest,
-          ),
-        {
-          label: `Move ${updates.length} layers`,
-          coalesceKey: `group-path-offset:${coalesceKey}`,
-        },
-      );
-      refreshDomEditGroupSelectionsFromPreview(domEditGroupSelectionsRef.current);
-    },
-    [
-      commitStudioManualEditManifestOptimistically,
-      domEditGroupSelectionsRef,
-      refreshDomEditGroupSelectionsFromPreview,
-    ],
-  );
-
-  const handleDomBoxSizeCommit = useCallback(
-    (selection: DomEditSelection, next: { width: number; height: number }) => {
-      commitStudioManualEditManifestOptimistically(
-        (manifest) => upsertStudioBoxSizeEdit(manifest, selection, next),
-        {
-          label: "Resize layer box",
-          coalesceKey: `box-size:${getDomEditTargetKey(selection)}`,
-        },
-      );
-      refreshDomEditSelectionFromPreview(selection);
-    },
-    [commitStudioManualEditManifestOptimistically, refreshDomEditSelectionFromPreview],
-  );
-
-  const handleDomRotationCommit = useCallback(
-    (selection: DomEditSelection, next: { angle: number }) => {
-      commitStudioManualEditManifestOptimistically(
-        (manifest) => upsertStudioRotationEdit(manifest, selection, next),
-        {
-          label: "Rotate layer",
-          coalesceKey: `rotation:${getDomEditTargetKey(selection)}`,
-        },
-      );
-      refreshDomEditSelectionFromPreview(selection);
-    },
-    [commitStudioManualEditManifestOptimistically, refreshDomEditSelectionFromPreview],
-  );
-
-  const handleDomManualEditsReset = useCallback(
-    (selection: DomEditSelection) => {
-      commitStudioManualEditManifestOptimistically(
-        (manifest) => removeStudioManualEditsForSelection(manifest, selection),
-        {
-          label: "Reset layer edits",
-          coalesceKey: `manual-reset:${getDomEditTargetKey(selection)}`,
-        },
-      );
-      applyCurrentStudioManualEditsToPreview(previewIframeRef.current);
-      refreshDomEditSelectionFromPreview(selection);
-    },
-    [
-      applyCurrentStudioManualEditsToPreview,
-      commitStudioManualEditManifestOptimistically,
-      refreshDomEditSelectionFromPreview,
-      previewIframeRef,
-    ],
-  );
-
-  const handleDomMotionCommit = useCallback(
-    (
-      selection: DomEditSelection,
-      motion: Omit<StudioGsapMotion, "kind" | "target" | "updatedAt">,
-    ) => {
-      commitStudioMotionManifestOptimistically(
-        (manifest) => upsertStudioGsapMotion(manifest, selection, motion),
-        {
-          label: "Set GSAP motion",
-          coalesceKey: `motion:${getDomEditTargetKey(selection)}`,
-        },
-      );
-      refreshDomEditSelectionFromPreview(selection);
-    },
-    [commitStudioMotionManifestOptimistically, refreshDomEditSelectionFromPreview],
-  );
-
-  const handleDomMotionClear = useCallback(
-    (selection: DomEditSelection) => {
-      commitStudioMotionManifestOptimistically(
-        (manifest) => removeStudioMotionForSelection(manifest, selection),
-        {
-          label: "Clear GSAP motion",
-          coalesceKey: `motion:${getDomEditTargetKey(selection)}`,
-        },
-      );
-      applyCurrentStudioMotionToPreview(previewIframeRef.current);
-      refreshDomEditSelectionFromPreview(selection);
-    },
-    [
-      applyCurrentStudioMotionToPreview,
-      commitStudioMotionManifestOptimistically,
-      refreshDomEditSelectionFromPreview,
-      previewIframeRef,
-    ],
-  );
-
-  const handleDomStyleCommit = useCallback(
-    async (property: string, value: string) => {
-      if (!domEditSelection) return;
-      if (isManualGeometryStyleProperty(property)) return;
-      if (!domEditSelection.capabilities.canEditStyles) return;
-      const importedFont = property === "font-family" ? resolveImportedFontAsset(value) : null;
-      const iframe = previewIframeRef.current;
-      const doc = iframe?.contentDocument;
-      if (doc) {
-        const el = findElementForSelection(doc, domEditSelection, activeCompPath);
-        if (el) {
-          el.style.setProperty(property, normalizeDomEditStyleValue(property, value));
-          if (property === "font-family") {
-            injectPreviewGoogleFont(doc, value);
-            if (importedFont) injectPreviewImportedFont(doc, importedFont);
-          }
-          if (property === "background-image" && isImageBackgroundValue(value)) {
-            el.style.setProperty("background-position", "center");
-            el.style.setProperty("background-repeat", "no-repeat");
-            el.style.setProperty("background-size", "contain");
-          }
-        }
-      }
-      const operations: PatchOperation[] = [
-        buildDomEditStylePatchOperation(property, normalizeDomEditStyleValue(property, value)),
-      ];
-      if (property === "background-image" && isImageBackgroundValue(value)) {
-        operations.push(
-          buildDomEditStylePatchOperation("background-position", "center"),
-          buildDomEditStylePatchOperation("background-repeat", "no-repeat"),
-          buildDomEditStylePatchOperation("background-size", "contain"),
-        );
-      }
-      try {
-        await persistDomEditOperations(domEditSelection, operations, {
-          label: "Edit layer style",
-          skipRefresh: true,
-          prepareContent: importedFont
-            ? (html, sourceFile) => ensureImportedFontFace(html, importedFont, sourceFile)
-            : undefined,
-        });
-      } catch (err) {
-        console.warn("[Studio] Style persist failed:", err instanceof Error ? err.message : err);
-      }
-      refreshDomEditSelectionFromPreview(domEditSelection);
-    },
-    [
-      activeCompPath,
-      domEditSelection,
-      persistDomEditOperations,
-      refreshDomEditSelectionFromPreview,
-      resolveImportedFontAsset,
-      previewIframeRef,
-    ],
-  );
-
-  const handleDomTextCommit = useCallback(
-    async (value: string, fieldKey?: string) => {
-      if (!domEditSelection) return;
-      if (!isTextEditableSelection(domEditSelection)) return;
-      const commitVersion = domTextCommitVersionRef.current + 1;
-      domTextCommitVersionRef.current = commitVersion;
-      const nextTextFields =
-        domEditSelection.textFields.length > 0
-          ? domEditSelection.textFields.map((field) =>
-              field.key === fieldKey ? { ...field, value } : field,
-            )
-          : [];
-      const nextContent =
-        nextTextFields.length > 1 || nextTextFields.some((field) => field.source === "child")
-          ? serializeDomEditTextFields(nextTextFields)
-          : value;
-      const iframe = previewIframeRef.current;
-      const doc = iframe?.contentDocument;
-      if (doc) {
-        const el = findElementForSelection(doc, domEditSelection, activeCompPath);
-        if (el) {
-          if (
-            nextTextFields.length > 1 ||
-            nextTextFields.some((field) => field.source === "child")
-          ) {
-            el.innerHTML = nextContent;
-          } else {
-            el.textContent = value;
-          }
-        }
-      }
-      await persistDomEditOperations(
-        domEditSelection,
-        [buildDomEditTextPatchOperation(nextContent)],
-        {
-          label: "Edit text",
-          skipRefresh: true,
-          shouldSave: () => domTextCommitVersionRef.current === commitVersion,
-        },
-      );
-      if (domTextCommitVersionRef.current !== commitVersion) return;
-
-      if (doc) {
-        const refreshed = findElementForSelection(doc, domEditSelection, activeCompPath);
-        if (refreshed) {
-          const nextSelection = buildDomSelectionFromTarget(refreshed);
-          if (nextSelection) {
-            applyDomSelection(nextSelection, { revealPanel: false, preserveGroup: true });
-          }
-        }
-      }
-    },
-    [
-      activeCompPath,
-      applyDomSelection,
-      buildDomSelectionFromTarget,
-      domEditSelection,
-      persistDomEditOperations,
-      previewIframeRef,
-    ],
-  );
-
-  const commitDomTextFields = useCallback(
-    async (
-      selection: DomEditSelection,
-      nextTextFields: DomEditTextField[],
-      options?: { importedFont?: ImportedFontAsset | null },
-    ) => {
-      const nextContent =
-        nextTextFields.length > 1 || nextTextFields.some((field) => field.source === "child")
-          ? serializeDomEditTextFields(nextTextFields)
-          : (nextTextFields[0]?.value ?? "");
-
-      const iframe = previewIframeRef.current;
-      const doc = iframe?.contentDocument;
-      if (doc) {
-        const el = findElementForSelection(doc, selection, activeCompPath);
-        if (el) {
-          if (
-            nextTextFields.length > 1 ||
-            nextTextFields.some((field) => field.source === "child")
-          ) {
-            el.innerHTML = nextContent;
-          } else {
-            el.textContent = nextContent;
-          }
-        }
-      }
-
-      const importedFont = options?.importedFont ?? null;
-      await persistDomEditOperations(selection, [buildDomEditTextPatchOperation(nextContent)], {
-        label: "Edit text",
-        skipRefresh: true,
-        prepareContent: importedFont
-          ? (html, sourceFile) => ensureImportedFontFace(html, importedFont, sourceFile)
-          : undefined,
-      });
-
-      if (doc) {
-        const refreshed = findElementForSelection(doc, selection, activeCompPath);
-        if (refreshed) {
-          const nextSelection = buildDomSelectionFromTarget(refreshed);
-          if (nextSelection) {
-            applyDomSelection(nextSelection, { revealPanel: false, preserveGroup: true });
-          }
-        }
-      }
-    },
-    [
-      activeCompPath,
-      applyDomSelection,
-      buildDomSelectionFromTarget,
-      persistDomEditOperations,
-      previewIframeRef,
-    ],
-  );
-
-  const handleDomTextFieldStyleCommit = useCallback(
-    async (fieldKey: string, property: string, value: string) => {
-      if (!domEditSelection) return;
-      const field = domEditSelection.textFields.find((entry) => entry.key === fieldKey);
-      if (!field) return;
-
-      if (field.source === "self") {
-        await handleDomStyleCommit(property, value);
-        return;
-      }
-
-      const normalizedValue = normalizeDomEditStyleValue(property, value);
-      const importedFont = property === "font-family" ? resolveImportedFontAsset(value) : null;
-      if (property === "font-family") {
-        const doc = previewIframeRef.current?.contentDocument;
-        if (doc) {
-          injectPreviewGoogleFont(doc, normalizedValue);
-          if (importedFont) injectPreviewImportedFont(doc, importedFont);
-        }
-      }
-      const nextTextFields = domEditSelection.textFields.map((entry) =>
-        entry.key === fieldKey
-          ? {
-              ...entry,
-              inlineStyles: {
-                ...entry.inlineStyles,
-                [property]: normalizedValue,
-              },
-              computedStyles: {
-                ...entry.computedStyles,
-                [property]: normalizedValue,
-              },
-            }
-          : entry,
-      );
-
-      await commitDomTextFields(domEditSelection, nextTextFields, { importedFont });
-    },
-    [
-      commitDomTextFields,
-      domEditSelection,
-      handleDomStyleCommit,
-      resolveImportedFontAsset,
-      previewIframeRef,
-    ],
-  );
-
-  const handleDomAddTextField = useCallback(
-    async (afterFieldKey?: string) => {
-      if (!domEditSelection) return null;
-      if (!domEditSelection.textFields.some((field) => field.source === "child")) return null;
-
-      const insertionIndex = domEditSelection.textFields.findIndex(
-        (field) => field.key === afterFieldKey,
-      );
-      const baseField =
-        domEditSelection.textFields[insertionIndex >= 0 ? insertionIndex : 0] ??
-        domEditSelection.textFields[0];
-      const nextField = buildDefaultDomEditTextField(baseField);
-      const nextTextFields = [...domEditSelection.textFields];
-      nextTextFields.splice(
-        insertionIndex >= 0 ? insertionIndex + 1 : nextTextFields.length,
-        0,
-        nextField,
-      );
-
-      await commitDomTextFields(domEditSelection, nextTextFields);
-      return nextField.key;
-    },
-    [commitDomTextFields, domEditSelection],
-  );
-
-  const handleDomRemoveTextField = useCallback(
-    async (fieldKey: string) => {
-      if (!domEditSelection) return;
-      const field = domEditSelection.textFields.find((entry) => entry.key === fieldKey);
-      if (!field) return;
-
-      if (field.source === "self") {
-        await handleDomTextCommit("", fieldKey);
-        return;
-      }
-
-      const nextTextFields = domEditSelection.textFields.filter((entry) => entry.key !== fieldKey);
-      await commitDomTextFields(domEditSelection, nextTextFields);
-    },
-    [commitDomTextFields, domEditSelection, handleDomTextCommit],
-  );
+  const {
+    resolveImportedFontAsset,
+    handleDomStyleCommit,
+    handleDomTextCommit,
+    handleDomTextFieldStyleCommit,
+    handleDomAddTextField,
+    handleDomRemoveTextField,
+    handleDomPathOffsetCommit,
+    handleDomGroupPathOffsetCommit,
+    handleDomBoxSizeCommit,
+    handleDomRotationCommit,
+    handleDomManualEditsReset,
+    handleDomMotionCommit,
+    handleDomMotionClear,
+    handleDomEditElementDelete,
+  } = useDomEditCommits({
+    activeCompPath,
+    previewIframeRef,
+    showToast,
+    commitStudioManualEditManifestOptimistically,
+    commitStudioMotionManifestOptimistically,
+    applyCurrentStudioManualEditsToPreview,
+    applyCurrentStudioMotionToPreview,
+    writeProjectFile,
+    domEditSaveTimestampRef,
+    editHistory,
+    fileTree,
+    importedFontAssetsRef,
+    projectId,
+    projectIdRef,
+    setRefreshKey,
+    domEditSelection,
+    domEditSelectionRef,
+    domEditGroupSelectionsRef,
+    applyDomSelection,
+    clearDomSelection,
+    refreshDomEditSelectionFromPreview,
+    refreshDomEditGroupSelectionsFromPreview,
+    buildDomSelectionFromTarget,
+  });
 
   // ── Effects ──
 
-  // Sync selection from preview document (the big one with attachErrorCapture)
+  // Sync selection from preview document on load / refresh
   // eslint-disable-next-line no-restricted-syntax
   useEffect(() => {
     if (!previewIframe) return;
@@ -772,84 +292,6 @@ export function useDomEditSession({
     applyStudioManualEditsToPreviewRef,
     applyStudioMotionToPreviewRef,
   ]);
-
-  const handleDomEditElementDelete = useCallback(
-    async (selection: DomEditSelection) => {
-      const pid = projectIdRef.current;
-      if (!pid) return;
-      const label = selection.label || selection.id || selection.selector || selection.tagName;
-      if (!confirmElementDelete(label, "element")) return;
-
-      const targetPath = selection.sourceFile || activeCompPath || "index.html";
-      try {
-        const response = await fetch(
-          `/api/projects/${pid}/files/${encodeURIComponent(targetPath)}`,
-        );
-        if (!response.ok) throw new Error(`Failed to read ${targetPath}`);
-
-        const data = (await response.json()) as { content?: string };
-        const originalContent = data.content;
-        if (typeof originalContent !== "string")
-          throw new Error(`Missing file contents for ${targetPath}`);
-
-        const patchTarget: { id?: string; selector?: string; selectorIndex?: number } = selection.id
-          ? {
-              id: selection.id,
-              selector: selection.selector,
-              selectorIndex: selection.selectorIndex,
-            }
-          : selection.selector
-            ? { selector: selection.selector, selectorIndex: selection.selectorIndex }
-            : ({} as never);
-        if (!patchTarget.id && !patchTarget.selector) {
-          throw new Error("Selected element has no patchable target");
-        }
-
-        const removeResponse = await fetch(
-          `/api/projects/${pid}/file-mutations/remove-element/${encodeURIComponent(targetPath)}`,
-          {
-            method: "POST",
-            headers: { "Content-Type": "application/json" },
-            body: JSON.stringify({ target: patchTarget }),
-          },
-        );
-        if (!removeResponse.ok) throw new Error(`Failed to delete element from ${targetPath}`);
-
-        const removeData = (await removeResponse.json()) as { changed?: boolean; content?: string };
-        const patchedContent =
-          typeof removeData.content === "string" ? removeData.content : originalContent;
-
-        domEditSaveTimestampRef.current = Date.now();
-        await saveProjectFilesWithHistory({
-          projectId: pid,
-          label: "Delete element",
-          kind: "timeline",
-          files: { [targetPath]: patchedContent },
-          readFile: async () => originalContent,
-          writeFile: writeProjectFile,
-          recordEdit: editHistory.recordEdit,
-        });
-
-        clearDomSelection();
-        usePlayerStore.getState().setSelectedElementId(null);
-        setRefreshKey((k) => k + 1);
-        showToast(`Deleted ${label}. Use Undo to restore it.`, "info");
-      } catch (error) {
-        const message = error instanceof Error ? error.message : "Failed to delete element";
-        showToast(message);
-      }
-    },
-    [
-      activeCompPath,
-      clearDomSelection,
-      domEditSaveTimestampRef,
-      editHistory.recordEdit,
-      projectIdRef,
-      setRefreshKey,
-      showToast,
-      writeProjectFile,
-    ],
-  );
 
   return {
     // State
